refactor(tv video rss): extract function to utils
